Use GenericCachingAllocator for pinned host memory
